Refactor QC classes
