+++ WebApp/src/uk/ac/exeter/QuinCe/data/QC/NoDataQCRecord.java
@@ -1,4 +1,4 @@
-package uk.ac.exeter.QuinCe.data;
+package uk.ac.exeter.QuinCe.data.QC;
 
 import java.util.List;
