+++ WebApp/src/uk/ac/exeter/QuinCe/web/files/DataScreenBean.java
@@ -13,9 +13,9 @@ import uk.ac.exeter.QuinCe.data.FileInfo;
 import uk.ac.exeter.QuinCe.data.RunType;
 import uk.ac.exeter.QuinCe.data.Instrument.Instrument;
 import uk.ac.exeter.QuinCe.data.Instrument.InstrumentDB;
+import uk.ac.exeter.QuinCe.data.QC.QCDB;
 import uk.ac.exeter.QuinCe.database.DatabaseException;
 import uk.ac.exeter.QuinCe.database.RecordNotFoundException;
-import uk.ac.exeter.QuinCe.database.QC.QCDB;
 import uk.ac.exeter.QuinCe.database.files.CommentSet;
 import uk.ac.exeter.QuinCe.database.files.CommentSetEntry;
 import uk.ac.exeter.QuinCe.database.files.DataFileDB;
